+++ minuet/Paragon.Runtime/WPF/Download/DownloadControl.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="Paragon.Runtime.WPF.DownloadControl"
+﻿<UserControl x:Class="Paragon.Runtime.WPF.Download.DownloadControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -37,7 +37,7 @@
                 <Setter.Value>
                     <ControlTemplate TargetType="{x:Type Button}">
                         <Grid Background="Transparent" HorizontalAlignment="Stretch" VerticalAlignment="Stretch">
-                            <luna:ButtonChrome x:Name="chrome" 
+                            <luna:ButtonChrome x:Name="chrome"
 							BorderBrush="{TemplateBinding BorderBrush}"
 							Fill="{TemplateBinding Background}"
 							HorizontalAlignment="Stretch" 
@@ -118,8 +118,8 @@
 									Style="{DynamicResource lunaMetallicDropDownStyle}" 
 									HorizontalAlignment="Stretch" 
 									Grid.Column="1"
-									Opacity="1">
-
+									Opacity="1"
+                                    >
                                         <Path Data="M0,0L3,3 6,0z" Margin="1,1,0,0" Stroke="{TemplateBinding Foreground}" Fill="{TemplateBinding Foreground}" />
                                     </Button>
                                 </Grid>
@@ -150,9 +150,9 @@
                                 <Setter Property="Visibility" TargetName="path" Value="Collapsed"/>
                                 <Setter Property="Grid.ColumnSpan" TargetName="Bd" Value="2"/>
                             </Trigger>
-                            <!--<Trigger Property="IsMouseOver" SourceName="PART_DropDown" Value="true">
+                            <Trigger Property="IsMouseOver" SourceName="PART_DropDown" Value="true">
                                 <Setter Property="Opacity" TargetName="PART_DropDown" Value="1"/>
-                            </Trigger>-->
+                            </Trigger>
                             <Trigger Property="IsEnabled" Value="false">
                                 <Setter Property="Foreground" Value="#ADADAD"/>
                             </Trigger>
@@ -168,7 +168,8 @@
             <ColumnDefinition Width="20"/>
         </Grid.ColumnDefinitions>
         <StackPanel Orientation="Horizontal" x:Name="itemsPanel" Grid.Row="0">
-            <!--<s:SplitButton Margin="2" 
+            <!--
+            <s:SplitButton Margin="2" 
                 Style="{StaticResource lunaMetallicSplitButtonStyle}">
                 <MenuItem Header="Open"/>
                 <MenuItem Header="Show In Folder"/>
@@ -179,21 +180,23 @@
                         <TextBlock Text="1 Byte"/>
                     </StackPanel>
                 </s:SplitButton.Content>
-            </s:SplitButton>-->
-            <!--<Button Margin="2" Background="Transparent" BorderBrush="Transparent" >
+            </s:SplitButton>
+            -->
+            <!--
+            <Button Margin="2" Background="Transparent" BorderBrush="Transparent" >
                 <Border BorderBrush="AliceBlue" BorderThickness="0" CornerRadius="5" Background="Red">
                     <StackPanel Margin="5,0,0,0" Width="150" Orientation="Vertical">
                         <TextBlock Text="filename" TextTrimming="CharacterEllipsis"/>
                         <TextBlock Text="1 Byte"/>
                     </StackPanel>
                 </Border>
-            </Button>-->
+            </Button>
+            -->
         </StackPanel>
 
-        <Button x:Name="closeButton" 
+        <Button x:Name="closeButton" Click="closeButtonClicked"
                 BorderBrush="Transparent"
-                HorizontalAlignment="Stretch"
                 VerticalContentAlignment="Center"
-                Grid.Column="1">X</Button>
+                Grid.Column="1" Content="x"/>
     </Grid>
 </UserControl>
